list files implementation create impl crate for proc_macro main crate reexports macro add utils.rs symlinked in both crates for code reuse export get_files method in main crate for use when in debug
