+++ impl/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "rust-embed-impl"
+version = "4.0.0"
+description = "Rust Custom Derive Macro which loads files into the rust binary at compile time during release and loads the file from the fs during dev"
+readme = "readme.md"
+documentation = "https://docs.rs/rust-embed"
+repository = "https://github.com/pyros2097/rust-embed"
+license = "MIT"
+keywords = ["http", "rocket", "static", "web", "server"]
+categories = ["web-programming::http-server"]
+authors = ["pyros2097 <pyros2097@gmail.com>"]
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = "0.11"
+quote = "0.3"
+walkdir = "2.1.4"
